Migrate base types
