--- src/components/network-map-viewer/utils/equipment-types.ts
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2023, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-export const EQUIPMENT_INFOS_TYPES = {
-    LIST: { type: 'LIST' },
-    MAP: { type: 'MAP' },
-    FORM: { type: 'FORM' },
-    TAB: { type: 'TAB' },
-    TOOLTIP: { type: 'TOOLTIP' },
-};
-
-export const EQUIPMENT_TYPES = {
-    SUBSTATION: 'SUBSTATION',
-    VOLTAGE_LEVEL: 'VOLTAGE_LEVEL',
-    LINE: 'LINE',
-    TWO_WINDINGS_TRANSFORMER: 'TWO_WINDINGS_TRANSFORMER',
-    THREE_WINDINGS_TRANSFORMER: 'THREE_WINDINGS_TRANSFORMER',
-    HVDC_LINE: 'HVDC_LINE',
-    GENERATOR: 'GENERATOR',
-    BATTERY: 'BATTERY',
-    LOAD: 'LOAD',
-    SHUNT_COMPENSATOR: 'SHUNT_COMPENSATOR',
-    TIE_LINE: 'TIE_LINE',
-    DANGLING_LINE: 'DANGLING_LINE',
-    STATIC_VAR_COMPENSATOR: 'STATIC_VAR_COMPENSATOR',
-    HVDC_CONVERTER_STATION: 'HVDC_CONVERTER_STATION',
-    VSC_CONVERTER_STATION: 'VSC_CONVERTER_STATION',
-    LCC_CONVERTER_STATION: 'LCC_CONVERTER_STATION',
-    SWITCH: 'SWITCH',
-};
