+++ src/index.ts
@@ -5,32 +5,64 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-export { NetworkAreaDiagramViewer } from './components/network-area-diagram-viewer/network-area-diagram-viewer';
-export type {
-    OnMoveNodeCallbackType,
-    OnMoveTextNodeCallbackType,
-    OnSelectNodeCallbackType,
-    OnToggleNadHoverCallbackType,
+export {
+    NetworkAreaDiagramViewer,
+    type OnMoveNodeCallbackType,
+    type OnMoveTextNodeCallbackType,
+    type OnSelectNodeCallbackType,
+    type OnToggleNadHoverCallbackType,
 } from './components/network-area-diagram-viewer/network-area-diagram-viewer';
-export type { DiagramMetadata } from './components/network-area-diagram-viewer/diagram-metadata';
-export { THRESHOLD_STATUS } from './components/network-area-diagram-viewer/dynamic-css-utils';
-export type { CSS_DECLARATION, CSS_RULE } from './components/network-area-diagram-viewer/dynamic-css-utils';
-export { SingleLineDiagramViewer } from './components/single-line-diagram-viewer/single-line-diagram-viewer';
 export type {
-    OnToggleSldHoverCallbackType,
-    OnBreakerCallbackType,
-    OnFeederCallbackType,
-    OnNextVoltageCallbackType,
-    OnBusCallbackType,
-    SLDMetadataComponent,
-    SLDMetadataComponentSize,
-    SLDMetadata,
-    SLDMetadataNode,
+    BusNodeMetadata,
+    DiagramMetadata,
+    EdgeMetadata,
+    LayoutParametersMetadata,
+    NodeMetadata,
+    SvgParametersMetadata,
+    TextNodeMetadata,
+} from './components/network-area-diagram-viewer/diagram-metadata';
+export {
+    type CSS_DECLARATION,
+    type CSS_RULE,
+    THRESHOLD_STATUS,
+} from './components/network-area-diagram-viewer/dynamic-css-utils';
+
+export {
+    SingleLineDiagramViewer,
+    type OnBreakerCallbackType,
+    type OnBusCallbackType,
+    type OnFeederCallbackType,
+    type OnNextVoltageCallbackType,
+    type OnToggleSldHoverCallbackType,
+    type SLDMetadata,
+    type SLDMetadataComponent,
+    type SLDMetadataComponentSize,
+    type SLDMetadataNode,
 } from './components/single-line-diagram-viewer/single-line-diagram-viewer';
+
 export { GeoData } from './components/network-map-viewer/network/geo-data';
 export { LineFlowMode, LineFlowColorMode } from './components/network-map-viewer/network/line-layer';
 export { MapEquipments } from './components/network-map-viewer/network/map-equipments';
-export { default as NetworkMap, DRAW_EVENT } from './components/network-map-viewer/network/network-map';
-export type { NetworkMapRef } from './components/network-map-viewer/network/network-map';
-
+export {
+    default as NetworkMap,
+    DRAW_EVENT,
+    type NetworkMapRef,
+} from './components/network-map-viewer/network/network-map';
 export { DRAW_MODES } from './components/network-map-viewer/network/draw-control';
+
+export {
+    EQUIPMENT_TYPES,
+    type LonLat,
+    type MapAnyLine,
+    type MapAnyLineWithType,
+    type MapEquipment,
+    type MapHvdcLine,
+    type MapHvdcLineWithType,
+    type MapLine,
+    type MapLineWithType,
+    type MapSubstation,
+    type MapTieLine,
+    type MapTieLineWithType,
+    type MapVoltageLevel,
+} from './equipment-types';
+export { type Coordinate, Country } from './powsybl';
